% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biniLasso.R
\name{biniFit}
\alias{biniFit}
\title{Fit a GLM using detected optimal cut-points}
\usage{
biniFit(
  data,
  optCuts,
  y,
  family,
  col_cuts = "opt_cuts",
  col_x = "opt_cuts_id",
  covariate_type = "dummy"
)
}
\arguments{
\item{data}{a data frame}

\item{optCuts}{a matrix containing a vector of optimal cut-points (as a list) corresponding to each covariate (may be null for some X), and a column of covariate names.}

\item{y}{response variable. A matrix with two columns for a Cox family model, and a vector otherwise}

\item{family}{any glmnet family option is available.}

\item{col_cuts}{column name of optimal cut-points in optCuts}

\item{col_x}{column name of covariate names in optCuts}

\item{covariate_type}{a string indicating whether the fitted Cox model/GLM should include categorical covariates derived from the optimal cut-points ("factor"),
or should use the indicator variables resulted from the optimal cut-points ("dummy").}
}
\value{
data frame including created categorical covariates
}
\description{
Fit a GLM using detected optimal cut-points
}
\examples{

simData <- data.frame(t = rexp(1000),
                      event = rbinom(1000, 1, 0.6),
                      x1 = rnorm(1000),
                      x2 = rnorm(1000),
                      x3 = rnorm(1000))

simData_converted <-
  cumBinarizer(data = simData,
               cols = c("x1", "x2", "x3"),
               method = "quantile",
               n_bins = 30)

simData_sim_optCuts <-
 opt_cuts_finder(x = simData_converted$x,
                 y = survival::Surv(simData$t, simData$event),
                 method = "biniLasso",
                 family = "cox",
                 lasso_rule = "1se",
                 lasso_nfolds = 10,
                 penalty.factor = NULL,
                 cols = c("x1", "x2", "x3"),
                 x_cuts = simData_converted$x_cuts)

biniFit(data = simData_converted,
        optCuts = .,
        y = Surv(simData_converted$tte, simData_converted$vital_status),
        family = "cox",
        col_cuts = "opt_cuts",
        col_x = "opt_cuts_id")


}
