% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biniLasso.R
\name{cumBinarizer}
\alias{cumBinarizer}
\title{Convert Numeric to Factor}
\usage{
cumBinarizer(
  data,
  cols,
  method = "quantile",
  n_bins = NULL,
  cuts_list = NULL,
  categorical_df = FALSE
)
}
\arguments{
\item{data}{a data frame}

\item{cols}{a vector of numeric column names (characters).}

\item{method}{either "quantile" (default) for using quantiles of a given column as its cut-points, or "fixed" to used a user provided vector of cut-points. See cuts_list input.}

\item{n_bins}{an integer value specifying number of bins to convert the numeric columns. Only needed for quantile method.}

\item{cuts_list}{a list of cut-points corresponding to each numeric column in the same order as entries of cols input.}

\item{categorical_df}{logical, indicates whether the function should returns a dataframe containing categorical version of categorized continuous covariates.}
}
\value{
data_cat a data frame including original numeric columns as well as the converted categorical columns.

x a matrix including dummy variables corresponding to converted categorical variables.

x_cuts a list of cut-points corresponding to each entry of cols and in the same order.
}
\description{
Convert Numeric to Factor
}
\examples{
simData <- data.frame(t = rexp(1000),
                      event = rbinom(1000, 1, 0.6),
                      x1 = rnorm(1000),
                      x2 = rnorm(1000),
                      x3 = rnorm(1000))

simData_converted_1 <-
  cumBinarizer(data = simData,
             cols = c("x1", "x2", "x3"),
             method = "quantile",
             n_bins = 30)

simData_converted_2 <-
  cumBinarizer(data = simData,
             cols = c("x1", "x2", "x3"),
             method = "fixed",
             cuts_list = list(c(0, 0.5),
                              c(-0.5, 0, 0.5),
                              c(-0.5, -0.1, 0.05, 0.1, 0.5)))

}
