% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biniLasso.R
\name{cuts_finder}
\alias{cuts_finder}
\title{Find a fixed number of cut-points for each predictor}
\usage{
cuts_finder(
  x = X_mat_tmp,
  y = Surv(simData_mod_sc1$Y, simData_mod_sc1$delta),
  lambda_grid = NULL,
  nCuts_per_group = 2,
  miniLasso = FALSE
)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars; each row is an observation vector. Can be in sparse matrix format (inherit from class "sparseMatrix" as in package Matrix). Requirement: nvars >1; in other words, x should have 2 or more columns.}

\item{y}{response variable. A matrix with two columns for a Cox family model, and a vector otherwise}

\item{lambda_grid}{a vector of sequence of lambda values. Can be left NULL.}

\item{nCuts_per_group}{integer, number of cut-points per predictor}

\item{miniLasso}{logical, whether to use miniLasso approach}
}
\value{
a vector of cut-points (character)
}
\description{
Find a fixed number of cut-points for each predictor
}
