% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biniLasso.R
\name{opt_fixed_nCuts}
\alias{opt_fixed_nCuts}
\title{Find optimal cut-points where number of cut-points per predictor is fixed}
\usage{
opt_fixed_nCuts(
  x,
  y,
  max_nCuts = 2,
  method = "biniLasso",
  family = "cox",
  lasso_rule = "1se",
  lasso_nfolds = 10,
  penalty.factor = NULL,
  lambda_grid = NULL,
  cols,
  x_cuts
)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars; each row is an observation vector. Can be in sparse matrix format (inherit from class "sparseMatrix" as in package Matrix). Requirement: nvars >1; in other words, x should have 2 or more columns.}

\item{y}{response variable. A matrix with two columns for a Cox family model, and a vector otherwise}

\item{max_nCuts}{integer, maximum number of cut-points allowed for each predictor}

\item{method}{specifies biniLasso vs. miniLasso, or "both".}

\item{family}{any glmnet family option is available.}

\item{lasso_rule}{Lasso rule for finding optimal lambda; either "1se" (default) or "min".}

\item{lasso_nfolds}{The number of lambda values - default is 10.}

\item{penalty.factor}{Separate penalty factors can be applied to each coefficient. This is a number that multiplies lasso lambda to allow differential shrinkage. Can be 0 for some variables, which implies no shrinkage, and that variable is always included in the model. Default is 1 for all variables . Also, any penalty.factor that is set to inf is converted to an exclude, and then internally reset to 1. Note: the penalty factors are internally rescaled to sum to nvars, and the lasso lambda sequence will reflect this change.}

\item{lambda_grid}{a vector of sequence of lambda values. Can be left NULL.}

\item{cols}{a vector of numeric column names (characters).}

\item{x_cuts}{a list of cut-points corresponding to each entry of cols and in the same order.}
}
\value{
a data.frame in which, each row corresponds to a method, and a column-list containing optimila cut-points per predictor
}
\description{
Find optimal cut-points where number of cut-points per predictor is fixed
}
\examples{


simData <- data.frame(t = rexp(1000),
                      event = rbinom(1000, 1, 0.6),
                      x1 = rnorm(1000),
                      x2 = rnorm(1000),
                      x3 = rnorm(1000))

simData_converted <-
  cumBinarizer(data = simData,
               cols = c("x1", "x2", "x3"),
               method = "quantile",
               n_bins = 30)

simData_sim_optCuts <-
 opt_cuts_finder(x = simData_converted$x,
                 y = survival::Surv(simData$t, simData$event),
                 method = "biniLasso",
                 family = "cox",
                 lasso_rule = "1se",
                 lasso_nfolds = 10,
                 penalty.factor = NULL,
                 cols = c("x1", "x2", "x3"),
                 x_cuts = simData_converted$x_cuts)

opt_fixed_nCuts(data = simData_converted,
                optCuts = .,
                y = Surv(simData_converted$tte, simData_converted$vital_status),
                family = "cox",
                max_nCuts = 2,
                method = "biniLasso",
                cols = c("x1", "x2", "x3"),
                x_cuts = simData_converted$x_cuts)



}
