% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biniLasso.R
\name{biniFit}
\alias{biniFit}
\title{Fit a GLM using detected optimal cut-points}
\usage{
biniFit(data, optCuts, y, family, col_cuts = "opt_cuts", col_x = "opt_cuts_id")
}
\arguments{
\item{data}{a data frame}

\item{optCuts}{a matrix containing a vector of optimal cut-points (as a list) corresponding to each covariate (may be null for some X), and a column of covariate names.}

\item{y}{response variable. A matrix with two columns for a Cox family model, and a vector otherwise}

\item{family}{any glmnet family option is available.}

\item{col_cuts}{column name of optimal cut-points in optCuts}

\item{col_x}{column name of covariate names in optCuts}
}
\value{
data a data frame including created categorical covariates
}
\description{
Fit a GLM using detected optimal cut-points
}
