% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biniLasso.R
\name{c_binarization}
\alias{c_binarization}
\title{Create cumulative binarized features}
\usage{
c_binarization(x, breaks, labels)
}
\arguments{
\item{x}{input numeric covariate for cumulative binarization}

\item{breaks}{breaking values to be used for binarization}

\item{labels}{labels to be used for the created cumulative binarized columns}
}
\value{
a matrix containing cumulative binarized features corresponds to x
}
\description{
Create cumulative binarized features
}
