% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biniLasso.R
\name{cuts_extractor}
\alias{cuts_extractor}
\title{Extract optimal cut-points from the fitted glmnet model.}
\usage{
cuts_extractor(glm_fit, cols, x_cuts)
}
\arguments{
\item{glm_fit}{the fitted glmnet object}

\item{cols}{a vector of numeric column names (characters)}

\item{x_cuts}{a list of cut-points corresponding to each entry of cols and in the same order.}
}
\value{
a list of optimal cut-points corresponding to the columns in cols and in the same order.
}
\description{
Extract optimal cut-points from the fitted glmnet model.
}
